GROK-13011: EDA | TS package: UI refactoring
Custom UI for PCA and PLS is moved to separate file.
