+++ packages/EDA/package.json
@@ -1,1 +1,47 @@
-{"name": "@datagrok/eda", "friendlyName": "EDATS", "version": "0.0.1", "description": "", "dependencies": {"datagrok-api": "latest", "cash-dom": "latest", "dayjs": "latest", "@datagrok-libraries/utils": "latest"}, "author": {"name": "Viktor Makarichev", "email": "vmakarichev@datagrok.ai"}, "devDependencies": {"webpack": "latest", "webpack-cli": "latest", "ts-loader": "latest", "typescript": "latest"}, "scripts": {"link-all": "", "debug-eda": "webpack && grok publish", "release-eda": "webpack && grok publish --release", "build-eda": "webpack", "build": "webpack", "debug-eda-dev": "webpack && grok publish dev", "release-eda-dev": "webpack && grok publish dev --release", "debug-eda-local": "webpack && grok publish local", "release-eda-local": "webpack && grok publish local --release"}, "canEdit": ["Developers"], "canView": ["All users"], "repository": {"type": "git", "url": "https://github.com/datagrok-ai/public.git", "directory": "packages/EDA"}, "sources": ["wasm/EDA.js"]}
+{
+    "name": "@datagrok/eda",
+    "friendlyName": "EDATS",
+    "version": "0.0.2",
+    "description": "",
+    "dependencies": {
+        "datagrok-api": "latest",
+        "cash-dom": "latest",
+        "dayjs": "latest",
+        "@datagrok-libraries/utils": "latest"
+    },
+    "author": {
+        "name": "Viktor Makarichev",
+        "email": "vmakarichev@datagrok.ai"
+    },
+    "devDependencies": {
+        "webpack": "latest",
+        "webpack-cli": "latest",
+        "ts-loader": "latest",
+        "typescript": "latest"
+    },
+    "scripts": {
+        "link-all": "",
+        "debug-eda": "webpack && grok publish",
+        "release-eda": "webpack && grok publish --release",
+        "build-eda": "webpack",
+        "build": "webpack",
+        "debug-eda-dev": "webpack && grok publish dev",
+        "release-eda-dev": "webpack && grok publish dev --release",
+        "debug-eda-local": "webpack && grok publish local",
+        "release-eda-local": "webpack && grok publish local --release"
+    },
+    "canEdit": [
+        "Developers"
+    ],
+    "canView": [
+        "All users"
+    ],
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/datagrok-ai/public.git",
+        "directory": "packages/EDA"
+    },
+    "sources": [
+        "wasm/EDA.js"
+    ]
+}
