+++ packages/EDA/src/utils.ts
@@ -27,35 +27,3 @@ export function checkComponenets(features: DG.ColumnList, components: number): v
   if (components > features.length)
     throw new Error(COMP_EXCESS);
 }
-
-/*
-// Check whether data processing requires much time (PCA, PLS)
-export function isProcExpensive(table, features, max) {
-  return ((table.rowCount > max) || (features.length > max));
-}
-
-// Check inputs of data for SVM testing generator
-export function checkGeneratorSVMinputs(samplesCount, featuresCount, min, max, violatorsPercentage) 
-{
-  if (min >= max)
-    throw new Error(INCORERRECT_MIN_MAX_MES);
-  
-  if (featuresCount < FEATURES_COUNT_MIN) 
-    throw new Error(INCORERRECT_FEATURES_MES);
-
-  if (samplesCount < SAMPLES_COUNT_MIN) 
-    throw new Error(INCORERRECT_SAMPLES_MES);
-
-  if ((violatorsPercentage < PERCENTAGE_MIN) || (violatorsPercentage > PERCENTAGE_MAX))
-    throw new Error(INCORERRECT_PERCENTAGE_MES);
-}
-
-// Check whether SVM computation requires much time
-export function isDataBigForSVM(featureCols, featuresMax, labelsCol, labelsMax) {
-  if (featureCols.length > featuresMax)
-    return true;
-
-  if (labelsCol.length > labelsMax)
-    return true;
-}
-*/
